refactoring - split into multiple modules replacement of variables also for imported files
