+++ lua/rest-nvim/utils/init.lua
@@ -3,6 +3,15 @@ math.randomseed(os.time())
 
 local M = {}
 
+-- go_to_line moves the cursor to the desired line in the provided buffer
+-- @param bufnr Buffer number, a.k.a id
+-- @param line the desired cursor position
+M.go_to_line = function (bufnr, line)
+  vim.api.nvim_buf_call(bufnr, function()
+    vim.fn.cursor(line, 1)
+  end)
+end
+
 M.uuid = function()
   local template = "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx"
   return string.gsub(template, "[xy]", function(c)
@@ -19,6 +28,17 @@ M.file_exists = function(file)
     file:close()
   end
   return file ~= nil
+end
+
+-- read_file Reads all lines from a file and returns the content as a table
+-- returns empty table if file does not exist
+M.read_file = function(file)
+  if not M.file_exists(file) then return {} end
+  local lines = {}
+  for line in io.lines(file) do
+    lines[#lines + 1] = line
+  end
+  return lines
 end
 
 -- read_env_file Reads the environment variables found in the `.env` file and
