+++ tests/external_file/user.json
@@ -1,4 +1,5 @@
 {
     "name": "morpheus",
-    "job": "leader"
+    "job": "leader",
+    "time" : "{{$timestamp}}"
 }
